refactored auth system
fixed require cycle
