+++ Contexts.js
@@ -1,0 +1,6 @@
+import { createContext } from 'react';
+
+export const AuthContext = createContext({
+  hasUser: false, 
+  setUser: () => {},
+});
